removed select geoplot
